Add subreddit hiding
One major and one minor issue: When a user's submissions is also hidden for the same as the subreddit hidden, two seperate "show" lines showup. Secondly, this is a pretty close copy of the code from hiding users, so a refactor (probably under common code at the top) should be done to avoid this repetition.